Complete implementation of clock constraints
Finished implementing clocks (very simple as the semantics are separate)
and clock constraints (with syntactic sugar for creating them).

Also decided to use the `portion` package from PyPI to use as intervals.
The library seems to be small and clean, and consists of all the needed
functionality for simulating PTA.

Next Steps
==========

- Implement the actual automaton.
  - I'll probably do an explicit representation, where each state is
    hashable and the transition function for each state is an explicit
    probabilistic/non-deterministic transition.